Criação dos métodos PATCH + Alterações

+++ src/app/main.py
@@ -70,6 +70,7 @@ Base.metadata.create_all(bind=engine)
 ### =============== Funções =============== ###
 #=============================================#
 
+''' #! Função descontinuada (Até o momento)
 #* Função de validar filtragens dentro do banco e retornar para a API
 def raise_http_exception(parameter, message: str):
     if not parameter:
@@ -77,7 +78,7 @@ def raise_http_exception(parameter, message: str):
             status_code=status.HTTP_404_NOT_FOUND, 
             detail=message
         )
-
+'''
 
 def get_db():
     db = SessionLocal() # Classe do Alchemy (Cria uma nova sessão do banco)
@@ -108,6 +109,12 @@ class UserSchema(BaseModel):
         #* Faz a manipulação de dados para JSON
         orm_mode = True
 
+#* Atualiza os dados dos usuários para a API: 
+class UserPatch(BaseModel):
+    name: Optional[str] = None
+    email: Optional[str] = None
+
+# ---
 
 #* Define o tipo dos dados pra Transaction:
 class TransactionCreate(BaseModel):
@@ -128,6 +135,13 @@ class TransactionSchema(BaseModel):
         #* Faz a manipulação de dados para JSON
         orm_mode = True
 
+#* Atualiza os dados da Transação para a API:
+class TransactionPatch(BaseModel):
+    description: Optional[str] = None
+    amount: Optional[float] = None
+    category_id: Optional[int] = None
+
+# ---
 
 #* Define os tipos de dados pra Category:
 class CategoryCreate(BaseModel):
@@ -142,6 +156,10 @@ class CategorySchema(BaseModel):
         #* Faz a manipulação de dados para JSON
         orm_mode = True
 
+#* Atualiza os dados das Categorias para a API:
+class CategoryPatch(BaseModel):
+    name: Optional[str] = None
+
 
 # ==================================================== #  
 ### =============== Endpoints da API =============== ###
@@ -154,16 +172,17 @@ class CategorySchema(BaseModel):
 @app.get("/api/users/", response_model=List[UserSchema])
 def get_user(db: Session = Depends(get_db)):
     users = db.query(User).all()
-    raise_http_exception(users, message="Users not founded")
+    if not users:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Users not found")
     return users
 
 #* Define o Endpoint para pegar as informações de 1 User (GET)
 @app.get("/api/users/{user_id}/", response_model=UserSchema)
 def get_user(user_id: int, db: Session = Depends(get_db)):
     user = db.query(User).filter(User.id == user_id).first()
-    raise_http_exception(user, message="User not found")
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
     return user
-
 
 #* Define o Endpoint da criação de User (POST)
 @app.post("/api/users/", response_model=UserSchema)
@@ -179,10 +198,28 @@ def create_user(user: UserCreate, db: Session = Depends(get_db)):
     db.refresh(db_user) # Atualiza o Usuario criado no Banco
     return db_user # Retorna o objeto para a FastAPI
 
+#* Define o Endpoint para atualizar User (PATCH)
+@app.patch("/api/users/{user_id}/", response_model=UserPatch)
+def update_user(user_id: int, user: UserPatch, db: Session = Depends(get_db)):
+    db_user = db.query(User).filter(User.id == user_id).first()
+
+    if db_user is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
+    
+    user_data = user.model_dump(exclude_unset=True, exclude_none=True) # Gera dicts apenas com as informações definidas na requisição
+
+    for key, value in user_data.items(): # Faz a alteração de todos os items que vieram com valores
+        setattr(db_user, key, value)
+    
+    db.commit() # Salva no banco
+    db.refresh(db_user) # Atualiza o 'User' no Banco
+    return db_user
+
 @app.delete("/api/users/{user_id}/")
 def delete_user(user_id: int, db: Session = Depends(get_db)):
     user = db.query(User).filter(User.id == user_id).first()
-    raise_http_exception(user, message="User not found")
+    if not user:
+        raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail="User not found")
     db.delete(user) # Deleta a informação
     db.commit() # Salva no banco
     return {
@@ -197,14 +234,16 @@ def delete_user(user_id: int, db: Session = Depends(get_db)):
 @app.get("/api/transactions/", response_model=List[TransactionSchema])
 def get_all_transactions(db: Session = Depends(get_db)):
     transactions = db.query(Transaction).all()
-    raise_http_exception(transactions, message="Transactions not founded")
+    if not transactions:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Transaction not found")
     return transactions
 
 #* Define o Endpoint para pegar as informações de 1 User (GET)
 @app.get("/api/users/{user_id}/transactions/", response_model=List[TransactionSchema])
 def get_transactions_for_user(user_id: int, db: Session = Depends(get_db)):
     user = db.query(User).filter(User.id == user_id).first() # Valida se o user_id é igual ao User.id do banco
-    raise_http_exception(user, message="User not found")
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
     
     transactions = db.query(Transaction).filter(Transaction.owner_id == user_id).all()
     # Valida se o transaction_id é igual ao Transaction.id do banco
@@ -215,10 +254,12 @@ def get_transactions_for_user(user_id: int, db: Session = Depends(get_db)):
 @app.post("/api/users/{user_id}/transactions/", response_model=TransactionSchema)
 def create_transaction(user_id: int, transaction: TransactionCreate, db: Session = Depends(get_db)):
     user_owner = db.query(User).filter(User.id == user_id).first()
-    raise_http_exception(user_owner, message="User not found")
+    if not user_owner:
+        raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail="User not found")
     
     category_exists = db.query(Category).filter(Category.id == transaction.category_id).first()
-    raise_http_exception(category_exists, message="Category not found")
+    if not category_exists:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Category not found")
 
     # Criando uma nova transação
     db_transaction = Transaction(
@@ -233,6 +274,27 @@ def create_transaction(user_id: int, transaction: TransactionCreate, db: Session
     db.refresh(db_transaction) # Atualiza a transação criada no Banco
     return db_transaction # Retorna o objeto para a FastAPI
 
+#* Define um Endpoint para atualização de uma transação de um usuário (PATCH)
+@app.patch("/api/users/{user_id}/transactions/{transaction_id}/", response_model=TransactionPatch)
+def update_transaction(user_id: int, transaction_id: int, transaction: TransactionPatch, db: Session = Depends(get_db)):
+    # Filtra a query do banco pelo id do usuario e id de transação igual ao do Banco.
+    db_transaction = db.query(Transaction).filter(
+        Transaction.owner_id == user_id, 
+        Transaction.id == transaction_id
+        ).first()
+    
+    if db_transaction is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Transaction not found")
+
+    transaction_data = transaction.model_dump(exclude_unset=True, exclude_none=True) # Gera dicts apenas com as informações definidas na requisição
+
+    for key, value in transaction_data.items(): # Faz a alteração de todos os items que vieram com valores
+        setattr(db_transaction, key, value)
+
+    db.commit() # Salva no banco
+    db.refresh(db_transaction) # Atualiza uma transação de um usuário no Banco
+    return db_transaction
+
 
 #* Define o Endpoint para deletar uma transação de um usuario (DELETE)
 @app.delete("/api/users/{user_id}/transactions/{transaction_id}/")
@@ -242,9 +304,11 @@ def delete_transaction(user_id:int, transaction_id:int, db: Session = Depends(ge
         Transaction.owner_id == user_id, 
         Transaction.id == transaction_id
         ).first()
-
-    raise_http_exception(del_transaction, 
-        message="Transaction not found or does not belong to the specified user."
+    
+    if not del_transaction:
+        raise HTTPException(
+            status_code=status.HTTP_409_CONFLICT,
+            detail="Transaction not found or does not belong to the specified user."
         )
     
     db.delete(del_transaction) # Deleta a informação
@@ -260,13 +324,16 @@ def delete_transaction(user_id:int, transaction_id:int, db: Session = Depends(ge
 ## === Endpoints para o recurso 'Categories' === ##
 #=================================================#
 
+#* Define o Endpoint para pegar todas categorias (GET)
 @app.get("/api/categories/", response_model=List[CategorySchema])
 def get_all_categories(db: Session = Depends(get_db)):
     categories = db.query(Category).all()
-    raise_http_exception(categories, message="Category not found")
+    if not categories:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Category not found")
 
     return categories
 
+#* Define o Endpoint para criar uma categoria (POST)
 @app.post("/api/categories/", response_model=CategorySchema)
 def create_category(category: CategoryCreate, db: Session = Depends(get_db)):
     # Valida se categoria ja existe no banco através do nome.
@@ -285,14 +352,30 @@ def create_category(category: CategoryCreate, db: Session = Depends(get_db)):
     db.refresh(db_category) # Atualiza a categoria dentro do banco 
     return db_category
 
+@app.patch("/api/categories/{category_id}/")
+def update_category(category_id: int, category: CategoryPatch, db: Session = Depends(get_db)):
+    db_category = db.query(Category).filter(Category.id == category_id).first()
+
+    if db_category is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Category not found")
+
+    category_data = category.model_dump(exclude_unset=True, exclude_none=True) # Gera dicts apenas com as informações definidas na requisição
+
+    for key, value in category_data.items(): # Faz a alteração de todos os items que vieram com valores
+        setattr(db_category, key, value)
+
+    db.commit() # Salva no banco
+    db.refresh(db_category) # Atualiza uma alteração de uma categoria no Banco
+    return db_category
+
+#* Define o Endpoint para deletar uma categoria (DELETE)
 @app.delete("/api/categories/{category_id}/")
 def delete_category(category_id: int, db: Session = Depends(get_db)):
     
     del_category = db.query(Category).filter(Category.id == category_id).first()
-    raise_http_exception(del_category, 
-        message="Category doesn't exist."
-        )
-    
+    if not del_category:
+        raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail="Category doesn't exist.")
+
     db.delete(del_category)
     db.commit()
